+++ CHANGELOG.md
@@ -5,8 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2021-02-11
+## [0.3.0]
+
+### Added
+
+- setColumn public method
+- routing by URL param "workspace"
+- workspace templates
+- default built-in configurstion of workspace
+- footer and heaader
+- set and get plugin config methods
+- ability to open panel in modal window
+
 ### Changed
+
+- EventSystem config object properties
+
+### Fixed
+
+- incorrect closing tag of "gridstack-content-container" div
+- bug with deleting panels from dashboard crash error
+
+## [0.2.0] - 2021-02-11
+
+### Changed
+
 - code source directory name to DTCD-WorkspaceSystem
 - downloading DTCD-SDK from [storage](http://storage.dev.isgneuro.com)
 - paths in source files to DTCD-SDK